# application/services/custom_services/alpha1/src/models/email.ts
import * as mongoose from 'mongoose';

const Schema = mongoose.Schema;

export const emailSchema = new Schema({
   name: String,
   email: String,
   phone_no: String
})

const emailModel = mongoose.model('email', emailSchema, 'email');
export default emailModel;